% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{CPULABenchmarks-package}
\alias{CPULABenchmarks}
\alias{CPULABenchmarks-package}
\title{CPULABenchmarks: An R package with 18 linear algebra operations for CPU benchmarking}
\description{
An R package with 18 linear algebra operations for CPU benchmarking
  all packed in a one-line function. Linear algebra operations follow Mauricio
  Vargas Sepulveda's `r-with-intel-mkl` GitHub repository and Dirk
  Eddelbuettel's `gcbd` R package.
}
\details{
A one-line function to run 18 basic benchmarks for linear algebra
operations run on the CPU. Benchmarks follow Mauricio Vargas Sepulveda's
    `r-with-intel-mkl` GitHub repository and Dirk Eddelbuettel's `gcbd`
R package.
}
\keyword{internal}
