% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run benchmarks}
\usage{
run(n = 100, which = NULL)
}
\arguments{
\item{n}{An integer with the number of times each benchmark will be run.}

\item{which}{A vector of characters with the codes of the tests to run, or
NULL to run them all.}
}
\value{
Print the summary, and also writes to disk the timing data frame and
  the summary itself.
}
\description{
Run benchmarks
}
\details{
**B01** Create, transpose, deform a 2500x2500 matrix
  **B02** Raise a 2400x2400 matrix to 1000th power
  **B03** Sort 7,000,000 random values
  **B04** Crossproduct of 2800x2800 matrix (b = a' * a)
  **B05** Solve linear regression over a 3000x3000 matrix (c = a \\ b')
  **B06** FFT over 2,400,000 random values
  **B07** Eigenvalues of a 640x640 random matrix
  **B08** Determinant of a 2500x2500 random matrix
  **B09** Cholesky decomposition of a 3000x3000 matrix
  **B10** Inverse of a 1600x1600 random matrix
  **B11** Find 3,500,000 Fibonacci numbers (vector calculation)
  **B12** Create a 3000x3000 Hilbert matrix (matrix calculation)
  **B13** Grand common divisors of 400,000 pairs (recursion)
  **C01** QR decomposition
  **C02** SVD decomposition
  **C03** LU decomposition
}
\examples{
\dontrun{
  run(1000, NULL)
}
}
